following apple rules, disabling izly on ios:
Bug Fix Submissions

The issues we've identified below are eligible to be resolved on your next update. If this submission includes bug fixes and you'd like to have it approved at this time, reply to this message and let us know. You do not need to resubmit your app for us to proceed.

Alternatively, if you'd like to resolve these issues now, please review the details, make the appropriate changes, and resubmit.

Guideline 5.1.1 - Legal - Privacy - Data Collection and Storage

The app does not meet all requirements for apps that offer highly regulated services or handle sensitive user data. Specifically:

- The account that submits the app must be enrolled in the Apple Developer Program as an organization, and not as an individual.

The guideline 5.1.1(ix) requirements give users confidence that apps operating in highly regulated fields or that require sensitive user information are qualified to provide these services and will responsibly manage their data.

Next Steps

To resolve this issue, it would be appropriate to take the following steps:

- The app must be submitted through an Apple Developer Program account enrolled as an organization. You may either enroll in a new Apple Developer Program account as an organization, or request that your individual account be converted to an organization account by contacting Apple Developer Support.

Please note that you cannot resolve this issue with documentation showing permission to publish this app on behalf of the content owner or institution.

Resources

- Learn how to transfer apps between accounts.
- Request a fee waiver for nonprofit organizations, accredited educational institutions, or government entities.
Please see attached screenshot for details.